refactor the code and introduce SafeProcessHandle.Start: Windows part (all tests passing)
