+++ src/libraries/System.Diagnostics.Process/src/System/Diagnostics/Process.Windows.cs
@@ -421,233 +421,6 @@ namespace System.Diagnostics
             }
         }
 
-        /// <summary>Starts the process using the supplied start info.</summary>
-        private unsafe bool StartWithCreateProcess(ProcessStartInfo startInfo, SafeFileHandle? stdinHandle, SafeFileHandle? stdoutHandle, SafeFileHandle? stderrHandle)
-        {
-            // See knowledge base article Q190351 for an explanation of the following code.  Noteworthy tricky points:
-            //    * The handles are duplicated as inheritable before they are passed to CreateProcess so
-            //      that the child process can use them
-
-            var commandLine = new ValueStringBuilder(stackalloc char[256]);
-            BuildCommandLine(startInfo, ref commandLine);
-
-            Interop.Kernel32.STARTUPINFO startupInfo = default;
-            Interop.Kernel32.PROCESS_INFORMATION processInfo = default;
-            Interop.Kernel32.SECURITY_ATTRIBUTES unused_SecAttrs = default;
-            SafeProcessHandle procSH = new SafeProcessHandle();
-
-            // Inheritable copies of the child handles for CreateProcess
-            SafeFileHandle? inheritableStdinHandle = null;
-            SafeFileHandle? inheritableStdoutHandle = null;
-            SafeFileHandle? inheritableStderrHandle = null;
-
-            // Take a global lock to synchronize all redirect pipe handle creations and CreateProcess
-            // calls. We do not want one process to inherit the handles created concurrently for another
-            // process, as that will impact the ownership and lifetimes of those handles now inherited
-            // into multiple child processes.
-
-            ProcessUtils.s_processStartLock.EnterWriteLock();
-            try
-            {
-                startupInfo.cb = sizeof(Interop.Kernel32.STARTUPINFO);
-
-                if (stdinHandle is not null || stdoutHandle is not null || stderrHandle is not null)
-                {
-                    Debug.Assert(stdinHandle is not null && stdoutHandle is not null && stderrHandle is not null, "All or none of the standard handles must be provided.");
-
-                    DuplicateAsInheritableIfNeeded(stdinHandle, ref inheritableStdinHandle);
-                    DuplicateAsInheritableIfNeeded(stdoutHandle, ref inheritableStdoutHandle);
-                    DuplicateAsInheritableIfNeeded(stderrHandle, ref inheritableStderrHandle);
-
-                    startupInfo.hStdInput = (inheritableStdinHandle ?? stdinHandle).DangerousGetHandle();
-                    startupInfo.hStdOutput = (inheritableStdoutHandle ?? stdoutHandle).DangerousGetHandle();
-                    startupInfo.hStdError = (inheritableStderrHandle ?? stderrHandle).DangerousGetHandle();
-
-                    // If STARTF_USESTDHANDLES is not set, the new process will inherit the standard handles.
-                    startupInfo.dwFlags = Interop.Advapi32.StartupInfoOptions.STARTF_USESTDHANDLES;
-                }
-
-                if (startInfo.WindowStyle != ProcessWindowStyle.Normal)
-                {
-                    startupInfo.wShowWindow = (short)GetShowWindowFromWindowStyle(startInfo.WindowStyle);
-                    startupInfo.dwFlags |= Interop.Advapi32.StartupInfoOptions.STARTF_USESHOWWINDOW;
-                }
-
-                // set up the creation flags parameter
-                int creationFlags = 0;
-                if (startInfo.CreateNoWindow) creationFlags |= Interop.Advapi32.StartupInfoOptions.CREATE_NO_WINDOW;
-                if (startInfo.CreateNewProcessGroup) creationFlags |= Interop.Advapi32.StartupInfoOptions.CREATE_NEW_PROCESS_GROUP;
-
-                // set up the environment block parameter
-                string? environmentBlock = null;
-                if (startInfo._environmentVariables != null)
-                {
-                    creationFlags |= Interop.Advapi32.StartupInfoOptions.CREATE_UNICODE_ENVIRONMENT;
-                    environmentBlock = GetEnvironmentVariablesBlock(startInfo._environmentVariables!);
-                }
-
-                string? workingDirectory = startInfo.WorkingDirectory;
-                if (workingDirectory.Length == 0)
-                {
-                    workingDirectory = null;
-                }
-
-                bool retVal;
-                int errorCode = 0;
-
-                if (startInfo.UserName.Length != 0)
-                {
-                    if (startInfo.Password != null && startInfo.PasswordInClearText != null)
-                    {
-                        throw new ArgumentException(SR.CantSetDuplicatePassword);
-                    }
-
-                    Interop.Advapi32.LogonFlags logonFlags = (Interop.Advapi32.LogonFlags)0;
-                    if (startInfo.LoadUserProfile && startInfo.UseCredentialsForNetworkingOnly)
-                    {
-                        throw new ArgumentException(SR.CantEnableConflictingLogonFlags, nameof(startInfo));
-                    }
-                    else if (startInfo.LoadUserProfile)
-                    {
-                        logonFlags = Interop.Advapi32.LogonFlags.LOGON_WITH_PROFILE;
-                    }
-                    else if (startInfo.UseCredentialsForNetworkingOnly)
-                    {
-                        logonFlags = Interop.Advapi32.LogonFlags.LOGON_NETCREDENTIALS_ONLY;
-                    }
-
-                    commandLine.NullTerminate();
-                    fixed (char* passwordInClearTextPtr = startInfo.PasswordInClearText ?? string.Empty)
-                    fixed (char* environmentBlockPtr = environmentBlock)
-                    fixed (char* commandLinePtr = &commandLine.GetPinnableReference())
-                    {
-                        IntPtr passwordPtr = (startInfo.Password != null) ?
-                            Marshal.SecureStringToGlobalAllocUnicode(startInfo.Password) : IntPtr.Zero;
-
-                        try
-                        {
-                            retVal = Interop.Advapi32.CreateProcessWithLogonW(
-                                startInfo.UserName,
-                                startInfo.Domain,
-                                (passwordPtr != IntPtr.Zero) ? passwordPtr : (IntPtr)passwordInClearTextPtr,
-                                logonFlags,
-                                null,            // we don't need this since all the info is in commandLine
-                                commandLinePtr,
-                                creationFlags,
-                                (IntPtr)environmentBlockPtr,
-                                workingDirectory,
-                                ref startupInfo,        // pointer to STARTUPINFO
-                                ref processInfo         // pointer to PROCESS_INFORMATION
-                            );
-                            if (!retVal)
-                                errorCode = Marshal.GetLastWin32Error();
-                        }
-                        finally
-                        {
-                            if (passwordPtr != IntPtr.Zero)
-                                Marshal.ZeroFreeGlobalAllocUnicode(passwordPtr);
-                        }
-                    }
-                }
-                else
-                {
-                    commandLine.NullTerminate();
-                    fixed (char* environmentBlockPtr = environmentBlock)
-                    fixed (char* commandLinePtr = &commandLine.GetPinnableReference())
-                    {
-                        retVal = Interop.Kernel32.CreateProcess(
-                            null,                // we don't need this since all the info is in commandLine
-                            commandLinePtr,      // pointer to the command line string
-                            ref unused_SecAttrs, // address to process security attributes, we don't need to inherit the handle
-                            ref unused_SecAttrs, // address to thread security attributes.
-                            true,                // handle inheritance flag
-                            creationFlags,       // creation flags
-                            (IntPtr)environmentBlockPtr, // pointer to new environment block
-                            workingDirectory,    // pointer to current directory name
-                            ref startupInfo,     // pointer to STARTUPINFO
-                            ref processInfo      // pointer to PROCESS_INFORMATION
-                        );
-                        if (!retVal)
-                            errorCode = Marshal.GetLastWin32Error();
-                    }
-                }
-
-                if (processInfo.hProcess != IntPtr.Zero && processInfo.hProcess != new IntPtr(-1))
-                    Marshal.InitHandle(procSH, processInfo.hProcess);
-                if (processInfo.hThread != IntPtr.Zero && processInfo.hThread != new IntPtr(-1))
-                    Interop.Kernel32.CloseHandle(processInfo.hThread);
-
-                if (!retVal)
-                {
-                    string nativeErrorMessage = errorCode == Interop.Errors.ERROR_BAD_EXE_FORMAT || errorCode == Interop.Errors.ERROR_EXE_MACHINE_TYPE_MISMATCH
-                        ? SR.InvalidApplication
-                        : GetErrorMessage(errorCode);
-
-                    throw CreateExceptionForErrorStartingProcess(nativeErrorMessage, errorCode, startInfo.FileName, workingDirectory);
-                }
-            }
-            catch
-            {
-                procSH.Dispose();
-                throw;
-            }
-            finally
-            {
-                // Only dispose duplicated handles, not the original handles passed by the caller.
-                // When the handle was invalid or already inheritable, no duplication was needed.
-                inheritableStdinHandle?.Dispose();
-                inheritableStdoutHandle?.Dispose();
-                inheritableStderrHandle?.Dispose();
-
-                ProcessUtils.s_processStartLock.ExitWriteLock();
-
-                commandLine.Dispose();
-            }
-
-            if (procSH.IsInvalid)
-            {
-                procSH.Dispose();
-                return false;
-            }
-
-            SetProcessHandle(procSH);
-            SetProcessId((int)processInfo.dwProcessId);
-            return true;
-        }
-
-        /// <summary>Duplicates a handle as inheritable if it's valid and not inheritable.</summary>
-        private static void DuplicateAsInheritableIfNeeded(SafeFileHandle sourceHandle, ref SafeFileHandle? duplicatedHandle)
-        {
-            // The user can't specify invalid handle via ProcessStartInfo.Standard*Handle APIs.
-            // However, Console.OpenStandard*Handle() can return INVALID_HANDLE_VALUE for a process
-            // that was started with INVALID_HANDLE_VALUE as given standard handle.
-            if (sourceHandle.IsInvalid)
-            {
-                return;
-            }
-
-            // When we know for sure that the handle is inheritable, we don't need to duplicate.
-            // When GetHandleInformation fails, we still attempt to call DuplicateHandle,
-            // just to keep throwing the same exception (backward compatibility).
-            if (Interop.Kernel32.GetHandleInformation(sourceHandle, out Interop.Kernel32.HandleFlags flags)
-                && (flags & Interop.Kernel32.HandleFlags.HANDLE_FLAG_INHERIT) != 0)
-            {
-                return;
-            }
-
-            IntPtr currentProcHandle = Interop.Kernel32.GetCurrentProcess();
-            if (!Interop.Kernel32.DuplicateHandle(currentProcHandle,
-                sourceHandle,
-                currentProcHandle,
-                out duplicatedHandle,
-                0,
-                bInheritHandle: true,
-                Interop.Kernel32.HandleOptions.DUPLICATE_SAME_ACCESS))
-            {
-                throw new Win32Exception(Marshal.GetLastWin32Error());
-            }
-        }
-
         private static ConsoleEncoding GetEncoding(int codePage)
         {
             Encoding enc = EncodingHelper.GetSupportedConsoleEncoding(codePage);
@@ -655,30 +428,6 @@ namespace System.Diagnostics
         }
 
         private bool _signaled;
-
-        private static void BuildCommandLine(ProcessStartInfo startInfo, ref ValueStringBuilder commandLine)
-        {
-            // Construct a StringBuilder with the appropriate command line
-            // to pass to CreateProcess.  If the filename isn't already
-            // in quotes, we quote it here.  This prevents some security
-            // problems (it specifies exactly which part of the string
-            // is the file to execute).
-            ReadOnlySpan<char> fileName = startInfo.FileName.AsSpan().Trim();
-            bool fileNameIsQuoted = fileName.StartsWith('"') && fileName.EndsWith('"');
-            if (!fileNameIsQuoted)
-            {
-                commandLine.Append('"');
-            }
-
-            commandLine.Append(fileName);
-
-            if (!fileNameIsQuoted)
-            {
-                commandLine.Append('"');
-            }
-
-            startInfo.AppendArgumentsTo(ref commandLine);
-        }
 
         /// <summary>Gets timing information for the current process.</summary>
         private ProcessThreadTimes GetProcessTimes()
@@ -792,35 +541,6 @@ namespace System.Diagnostics
         private static ConsoleEncoding GetStandardInputEncoding() => GetEncoding((int)Interop.Kernel32.GetConsoleCP());
 
         private static ConsoleEncoding GetStandardOutputEncoding() => GetEncoding((int)Interop.Kernel32.GetConsoleOutputCP());
-
-        private static string GetEnvironmentVariablesBlock(DictionaryWrapper sd)
-        {
-            // https://learn.microsoft.com/windows/win32/procthread/changing-environment-variables
-            // "All strings in the environment block must be sorted alphabetically by name. The sort is
-            //  case-insensitive, Unicode order, without regard to locale. Because the equal sign is a
-            //  separator, it must not be used in the name of an environment variable."
-
-            var keys = new string[sd.Count];
-            sd.Keys.CopyTo(keys, 0);
-            Array.Sort(keys, StringComparer.OrdinalIgnoreCase);
-
-            // Join the null-terminated "key=val\0" strings
-            var result = new StringBuilder(8 * keys.Length);
-            foreach (string key in keys)
-            {
-                string? value = sd[key];
-
-                // Ignore null values for consistency with Environment.SetEnvironmentVariable
-                if (value != null)
-                {
-                    result.Append(key).Append('=').Append(value).Append('\0');
-                }
-            }
-
-            return result.ToString();
-        }
-
-        private static string GetErrorMessage(int error) => Interop.Kernel32.GetMessage(error);
 
         /// <summary>Gets the friendly name of the process.</summary>
         public string ProcessName
