+++ src/libraries/System.Diagnostics.Process/src/System/Diagnostics/ProcessUtils.cs
@@ -1,6 +1,7 @@
 // Licensed to the .NET Foundation under one or more agreements.
 // The .NET Foundation licenses this file to you under the MIT license.
 
+using System.ComponentModel;
 using System.IO;
 using System.Threading;
 
@@ -9,6 +10,7 @@ namespace System.Diagnostics
     internal static partial class ProcessUtils
     {
         internal static readonly ReaderWriterLockSlim s_processStartLock = new ReaderWriterLockSlim();
+        internal static int s_cachedSerializationSwitch;
 
         internal static string? FindProgramInPath(string program)
         {
@@ -30,6 +32,13 @@ namespace System.Diagnostics
             }
 
             return null;
+        }
+
+        internal static Win32Exception CreateExceptionForErrorStartingProcess(string errorMessage, int errorCode, string fileName, string? workingDirectory)
+        {
+            string directoryForException = string.IsNullOrEmpty(workingDirectory) ? Directory.GetCurrentDirectory() : workingDirectory;
+            string msg = SR.Format(SR.ErrorStartingProcess, fileName, directoryForException, errorMessage);
+            return new Win32Exception(errorCode, msg);
         }
     }
 }
